admin/server/requirements.txt
flask
flask-pymongo
pymongo
python-dotenv
flask-cors
gunicorn
cloudinary
